ET gen Hackthone/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg[binary]==3.1.19
psycopg-pool==3.2.3
langgraph==0.2.50
langgraph-checkpoint-postgres==2.0.7
langchain-anthropic==0.3.0
langchain-core==0.3.0
langsmith==0.1.140
apscheduler==3.10.4
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.9.0
aiohttp==3.10.0
